move helper functions to separate files
